Cast time variable to int before scaling model input in SDXLPipeline for improved type consistency during processing.

+++ pipeline/pipeline.py
@@ -98,7 +98,8 @@ class SDXLPipeline:
             print(f"latents before scale_mode_input: shape={latents.shape}, dtype={latents.dtype}, device={latents.device}")
             print(f"latents before scale_mode_input | Mean: {latents.mean():.6f} | Std: {latents.std():.6f} | Sum: {latents.sum():.6f}")
 
-            latent_model_input = self.scheduler.scale_model_input(latents, t)
+            #cast t to int
+            latent_model_input = self.scheduler.scale_model_input(latents, t.to(torch.int32))
             print(f"latent_model_input: shape={latent_model_input.shape}, dtype={latent_model_input.dtype}, device={latent_model_input.device}")
             print(f"latent_model_input | Mean: {latent_model_input.mean():.6f} | Std: {latent_model_input.std():.6f} | Sum: {latent_model_input.sum():.6f}")
